Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-    double x,y;
+    float x,y;
     scanf("%f %f",&x,&y);
-    double z=x*y;
+    float z=x*y;
     printf("Product: %.2f",z);
     return 0;
 }
